# backend/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/rooms":{"post":{"tags":["room-controller"],"operationId":"createRoom","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoomRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminTokenDto"}}}}}}},"/api/rooms/{uuid}/users":{"post":{"tags":["room-controller"],"operationId":"addUser","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/rooms/{uuid}/admin":{"post":{"tags":["room-controller"],"operationId":"addAdmin","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"Authorization","in":"header","description":"Admin токен в формате 'AdminToken {uuid}'","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/rooms/{uuid}":{"get":{"tags":["room-controller"],"operationId":"getRoom","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomDto"}}}}}}}},"components":{"schemas":{"CreateRoomRequest":{"type":"object","properties":{"username":{"type":"string"},"uuid":{"type":"string","format":"uuid"}}},"AdminTokenDto":{"type":"object","properties":{"adminToken":{"type":"string","format":"uuid"}}},"CreateUserRequest":{"type":"object","properties":{"username":{"type":"string"}}},"RoomDto":{"type":"object","properties":{"roomUuid":{"type":"string","format":"uuid"},"status":{"type":"string"}}},"UserDto":{"type":"object","properties":{"name":{"type":"string"},"room":{"$ref":"#/components/schemas/RoomDto"},"admin":{"type":"boolean"}}}}}}